code organization cleanup
rename the class that encapsulated IHost and the embedded Kestrel
instance to EmbeddedWebServer

add comments explaining the various pieces
