+++ src/Microsoft.Diagnostics.NETCore.Client/WebSocketServer/IWebSocketStreamAdapter.cs
@@ -3,6 +3,8 @@
 
 namespace Microsoft.Diagnostics.NETCore.Client.WebSocketServer;
 
+// The streams returned by IWebSocketServer implement the usual .NET Stream class, but they also
+// expose a way to check if the underlying websocket connection is still open.
 public interface IWebSocketStreamAdapter
 {
     public bool IsConnected { get; }
