+++ src/Microsoft.Diagnostics.NETCore.Client/WebSocketServer/WebSocketServerFactory.cs
@@ -3,6 +3,8 @@ using Microsoft.Extensions.Logging;
 
 namespace Microsoft.Diagnostics.NETCore.Client.WebSocketServer;
 
+// This interface allows dotnet-dsrouter to install a callback that will create IWebSocketServer instances.
+// This is used to avoid a dependency on ASP.NET in the client library.
 internal class WebSocketServerFactory
 {
     internal static void SetBuilder(Func<IWebSocketServer> builder)
